pyproject.toml
[project]
name = "skin-guardian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "numpy>=2.4.1",
    "onnxruntime>=1.23.2",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]
